+++ engine/src/engine.cpp
@@ -137,40 +137,40 @@ void Engine::ChangeScene(Scene* s) {
   }
 }
 
-//void Scene::Update(const double& dt) { ents.update(dt); }
-//
-//void Scene::Render() { ents.render(); }
-//
-//bool Scene::isLoaded() const 
-//{
-//    std::lock_guard<std::mutex> lck(_loaded_mtx);
-//    // Are we already loading asynchronously?
-//    if (_loaded_future.valid() // yes
-//        &&                     // Has it finished?
-//        _loaded_future.wait_for(chrono::seconds(0)) ==
-//            future_status::ready) {
-//      // Yes
-//      _loaded_future.get();
-//		//getFromFuture(_loaded_future);
-//      _loaded = true;
-//    }
-//	//_loaded = true;
-//    return _loaded;
-//}
-//
-//void Scene::setLoaded(bool b) {
-//  {
-//    std::lock_guard<std::mutex> lck(_loaded_mtx);
-//    _loaded = b;
-//  }
-//}
-//
-//void Scene::UnLoad() {
-//  ents.list.clear();
-//  setLoaded(false);
-//}
-//
-//void Scene::LoadAsync() { _loaded_future = std::async(&Scene::Load, this); }
+void Scene::Update(const double& dt) { ents.update(dt); }
+
+void Scene::Render() { ents.render(); }
+
+bool Scene::isLoaded() const 
+{
+    std::lock_guard<std::mutex> lck(_loaded_mtx);
+    // Are we already loading asynchronously?
+    if (_loaded_future.valid() // yes
+        &&                     // Has it finished?
+        _loaded_future.wait_for(chrono::seconds(0)) ==
+            future_status::ready) {
+      // Yes
+      _loaded_future.get();
+		//getFromFuture(_loaded_future);
+      _loaded = true;
+    }
+	//_loaded = true;
+    return _loaded;
+}
+
+void Scene::setLoaded(bool b) {
+  {
+    std::lock_guard<std::mutex> lck(_loaded_mtx);
+    _loaded = b;
+  }
+}
+
+void Scene::UnLoad() {
+  ents.list.clear();
+  setLoaded(false);
+}
+
+void Scene::LoadAsync() { _loaded_future = std::async(&Scene::Load, this); }
 
 sf::Vector2u Engine::getWindowSize() { return _window->getSize(); }
 
@@ -193,4 +193,4 @@ long long last() {
 }
 } // namespace timing
 
-//Scene::~Scene() { UnLoad(); }
+Scene::~Scene() { UnLoad(); }
