+++ engine/src/engine.h
@@ -1,59 +1,59 @@
 #pragma once
 #include <SFML/Graphics/RenderWindow.hpp>
 #include <ecm.h>
-//#include <future>
+#include <future>
 #include <maths.h>
-//#include <mutex>
+#include <mutex>
 #include <string>
 
-//class Scene {
-//public:
-//  Scene() = default;
-//  virtual ~Scene();
-//  virtual void Load() = 0;
-//  virtual void LoadAsync();
-//  virtual void UnLoad();
-//  virtual void Update(const double& dt);
-//  virtual void Render();
-//  bool isLoaded() const;
-//  std::shared_ptr<Entity> makeEntity();
-//
-//  EntityManager ents;
-//
-//protected:
-//  void setLoaded(bool);
-//  mutable bool _loaded;
-//  mutable std::future<void> _loaded_future;
-//  mutable std::mutex _loaded_mtx;
-//  void getFromFuture(std::future<void>& fut)
-//  {
-//	  try
-//	  {
-//		  return fut.get();
-//	  }
-//	  catch (...)
-//	  {
-//		  fut = {};
-//		  throw;
-//	  }
-//  };
-//private:
-//  /*mutable bool _loaded;
-//  mutable std::future<void> _loaded_future;
-//  mutable std::mutex _loaded_mtx;
-//  void getFromFuture(std::future<void>& fut)
-//  {
-//	  try
-//	  {
-//		  return fut.get();
-//	  }
-//	  catch (...)
-//	  {
-//		  fut = {};
-//		  throw;
-//	  }
-//  };*/
-//};
+class Scene {
+public:
+  Scene() = default;
+  virtual ~Scene();
+  virtual void Load() = 0;
+  virtual void LoadAsync();
+  virtual void UnLoad();
+  virtual void Update(const double& dt);
+  virtual void Render();
+  bool isLoaded() const;
+  std::shared_ptr<Entity> makeEntity();
+
+  EntityManager ents;
+
+protected:
+  void setLoaded(bool);
+  mutable bool _loaded;
+  mutable std::future<void> _loaded_future;
+  mutable std::mutex _loaded_mtx;
+  void getFromFuture(std::future<void>& fut)
+  {
+	  try
+	  {
+		  return fut.get();
+	  }
+	  catch (...)
+	  {
+		  fut = {};
+		  throw;
+	  }
+  };
+private:
+  /*mutable bool _loaded;
+  mutable std::future<void> _loaded_future;
+  mutable std::mutex _loaded_mtx;
+  void getFromFuture(std::future<void>& fut)
+  {
+	  try
+	  {
+		  return fut.get();
+	  }
+	  catch (...)
+	  {
+		  fut = {};
+		  throw;
+	  }
+  };*/
+};
 
 class Engine {
 public:
